Refactored and cleaned up directories. Also added JsonNode to convert simulation results string to Json object. (Not yet tested).
Also added simulation input parameters to Simulation class to store the input values.
