+++ src/main/java/com/example/batterysimulator_db_operations/Clients/SimulationResultRepository.java
@@ -1,5 +1,6 @@
-package com.example.batterysimulator_db_operations;
+package com.example.batterysimulator_db_operations.Clients;
 
+import com.example.batterysimulator_db_operations.POJOs.SimulationResults;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 public interface SimulationResultRepository extends JpaRepository<SimulationResults, Long> {
