+++ src/main/java/com/example/batterysimulator_db_operations/SimulationController.java
@@ -1,5 +1,8 @@
 package com.example.batterysimulator_db_operations;
 
+import com.example.batterysimulator_db_operations.POJOs.Simulation;
+import com.fasterxml.jackson.databind.JsonNode;
+import com.fasterxml.jackson.databind.ObjectMapper;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
@@ -10,6 +13,7 @@ import java.util.Optional;
 public class SimulationController {
 
     private final SimulationService simulationService;
+    private Simulation simulation;
 
     public SimulationController(SimulationService simulationService) {
         this.simulationService = simulationService;
@@ -27,8 +31,18 @@ public class SimulationController {
     }
 
     @PostMapping("/createSimulation")
-    public ResponseEntity<String>create(@RequestBody Simulation sim) {
-        simulationService.saveSimulation(sim);
+    public ResponseEntity<String>create(@RequestBody String simResults) {
+
+        try {
+            ObjectMapper objectMapper = new ObjectMapper();
+            // Deserialize the JSON string to JsonNode
+            JsonNode jsonNode = objectMapper.readTree(simResults);
+            // Convert JsonNode to Simulation object
+            simulation = objectMapper.treeToValue(jsonNode, Simulation.class);
+        } catch (Exception e) {
+            e.printStackTrace();
+        }
+        simulationService.saveSimulation(simulation);
         return new ResponseEntity<>("Simulation created successfully", HttpStatus.OK);
     }
 }
